lab analysis

+++ go/mev_auction/bsc-rpc-client/main.go
@@ -8,9 +8,11 @@ import (
 )
 
 func main() {
-	err1 := os.Remove("./log/rpc.log")
-	if err1 != nil {
-		panic(err1)
+	if fileExists("./log/rpc.log") {
+		err1 := os.Remove("./log/rpc.log")
+		if err1 != nil {
+			panic(err1)
+		}
 	}
 	ctx := context.Background()
 	cancelCtx, cancel := context.WithCancel(ctx)
@@ -26,13 +28,32 @@ func main() {
 	if err != nil {
 		panic(err)
 	}
-	isBlock := true // 是否开启模拟搜索者阻塞
-	core.InitSearcher(cancelCtx, 1, isBlock)
-	core.InitUser(cancelCtx, 1)
 
-	time.Sleep(1 * time.Minute)
+	openBlock := false // 是否开启模拟搜索者阻塞
+	if openBlock {
+		go func() {
+			time.Sleep(1 * time.Minute)
+			core.GlobalIsBlock = true
+		}()
+	}
+	core.InitSearcher(cancelCtx, 2, openBlock)
+	core.InitUser(cancelCtx, 3)
+
+	time.Sleep(3 * time.Minute)
 
 	cancel() // 停止所有用户和搜索者
 
 	time.Sleep(10 * time.Second) // 再同步几个区块
+}
+
+func fileExists(path string) bool {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true // 文件存在
+	}
+	if os.IsNotExist(err) {
+		return false // 文件不存在
+	}
+	// 其他错误，例如权限问题，也认为文件存在与否不确定
+	return false
 }
